+++ Bidon/SDK/Banner/BannerAdManager.swift
@@ -74,6 +74,12 @@ final class BannerAdManager: NSObject {
         viewContext: AdViewContext,
         auctionKey: String?
     ) {
+        guard BidonSdk.isInitialized else {
+            Logger.warning("Bidon SDK is not initialized or failed initialization. Initialize SDK first")
+            delegate?.adManager(self, didFailToLoad: .message("SDK is not initialized"), auctionInfo: auctionInfo)
+            return
+        }
+
         guard state.isIdle else {
             Logger.warning("Banner ad manager is not idle. Loading attempt is prohibited.")
             return