+++ ThirdPartyMediationAdapters/AppLovinMediationBidonAdapter/AdInstance/FullscreenAdInstance.m
@@ -12,12 +12,14 @@
 
 - (instancetype)initWithEcpm:(double)ecpm
                     demandId:(nonnull NSString *)demandId
+                          ad:(id<BDNAd>)ad
                   adInstance:(id<BDNFullscreenAd>)adInstance {
     self = [super init];
     if (self) {
         _ecpm = ecpm;
         _demandId = demandId;
         _adInstance = adInstance;
+        _ad = ad;
     }
     return self;
 }
@@ -25,6 +27,7 @@
 - (id)copyWithZone:(NSZone *)zone {
     return [[FullscreenAdInstance alloc] initWithEcpm:self.ecpm
                                              demandId:self.demandId
+                                                   ad:self.ad
                                            adInstance:self.adInstance];
 }