+++ ThirdPartyMediationAdapters/ISBidonCustomAdapter/ISBidonCustomAdapter.h
@@ -1,0 +1,11 @@
+#import <Foundation/Foundation.h>
+
+//! Project version number for ISBidonCustomAdapter.
+FOUNDATION_EXPORT double ISBidonCustomAdapterVersionNumber;
+
+//! Project version string for ISBidonCustomAdapter.
+FOUNDATION_EXPORT const unsigned char ISBidonCustomAdapterVersionString[];
+
+// Public interface
+#import <ISBidonCustomAdapter/ISBidonCustomAdapter-Swift.h>
+