+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # develop
 
 ## New features
+- APDM-1305 BCALP Adapter to Bidon - Interstitial 
+- BDN-978 Initialization failure state
+- BDN-942 New Bidon SDK init logic
 - BDN-975 GitHub Actions
 
 ## Fixes